Replace $this->input, $this->session, and remove $this->load calls

+++ Modules/Core/src/Services/CustomFieldsService.php
@@ -108,9 +108,7 @@ class CustomFieldsService extends BaseService
      * @originalFile CustomField.php
      */
     public function getPositions($table_name = false)
-    {
-        $this->load->model(['custom_fields/mdl_client_custom', 'custom_fields/mdl_invoice_custom', 'custom_fields/mdl_payment_custom', 'custom_fields/mdl_quote_custom', 'custom_fields/mdl_user_custom']);
-        $p         = $table_name ? 'ip_' : '';
+    {$p         = $table_name ? 'ip_' : '';
         $s         = $table_name ? '_custom' : '';
         $positions = [$p . 'client' . $s => Mdl_client_custom::$positions, $p . 'invoice' . $s => Mdl_invoice_custom::$positions, $p . 'payment' . $s => Mdl_payment_custom::$positions, $p . 'quote' . $s => Mdl_quote_custom::$positions, $p . 'user' . $s => Mdl_user_custom::$positions];
         foreach ($positions as $key => $val) {
@@ -211,9 +209,7 @@ class CustomFieldsService extends BaseService
         if ( ! $this->used($id)) {
             $custom_field = $this->getById($id);
             // Remove MULTIPLE|SINGLE CHOICE values
-            if (preg_match('/CHOICE/', $custom_field->custom_field_type)) {
-                $this->load->model('custom_values/mdl_custom_values');
-                $this->mdl_custom_values->deleteAllFid($id);
+            if (preg_match('/CHOICE/', $custom_field->custom_field_type)) {$this->mdl_custom_values->deleteAllFid($id);
             }
             // Remove reference in custom table
             $base = strtr($custom_field->custom_field_table, ['ip_' => '']) . '_field';
@@ -260,9 +256,7 @@ class CustomFieldsService extends BaseService
      * @originalFile CustomField.php
      */
     public function getValueForField($field_id, $custom_field_model, $object)
-    {
-        $this->load->model('custom_fields/' . $custom_field_model);
-        $cf_table             = str_replace('mdl_', '', $custom_field_model);
+    {$cf_table             = str_replace('mdl_', '', $custom_field_model);
         $cf_model_name        = str_replace('_custom', '', $cf_table);
         $value                = $this->{$custom_field_model}->where($cf_table . '_fieldid', $field_id)->where($cf_model_name . '_id', $object->{$cf_model_name . '_id'})->get()->result();
         $value_key            = $cf_table . '_fieldvalue';
@@ -280,10 +274,7 @@ class CustomFieldsService extends BaseService
      * @originalFile CustomField.php
      */
     public function getValuesForFields($custom_field_model, $model_id)
-    {
-        $this->load->model('custom_fields/' . $custom_field_model);
-        $this->load->model('custom_values/mdl_custom_values');
-        $fields = $this->{$custom_field_model}->byId($model_id)->get()->result();
+    {$fields = $this->{$custom_field_model}->byId($model_id)->get()->result();
         if (empty($fields)) {
             return [];
         }
@@ -331,9 +322,7 @@ class CustomFieldsService extends BaseService
      * @originalFile CustomField.php
      */
     private function renameColumn($table_name, $old_column_name, $new_column_name)
-    {
-        $this->load->dbforge();
-        $column = [$old_column_name => ['name' => $new_column_name, 'type' => 'VARCHAR', 'constraint' => 50]];
+    {$column = [$old_column_name => ['name' => $new_column_name, 'type' => 'VARCHAR', 'constraint' => 50]];
         $this->dbforge->modify_column($table_name, $column);
     }
 
@@ -343,9 +332,7 @@ class CustomFieldsService extends BaseService
      * @originalFile CustomField.php
      */
     private function addColumn($table_name, $column_name)
-    {
-        $this->load->dbforge();
-        $column = [$column_name => ['type' => 'VARCHAR', 'constraint' => 256]];
+    {$column = [$column_name => ['type' => 'VARCHAR', 'constraint' => 256]];
         $this->dbforge->addColumn($table_name, $column);
     }
 }
